pyproject.toml
[project]
name = "mock-schedule-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.24",
    "redis[hiredis]>=7.4.0",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]



[tool.mypy]
# Enable the SQLAlchemy plugin here
plugins = ["sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true
